+++ docs/docs/.vitepress/config.ts
@@ -18,7 +18,7 @@ export default defineConfig({
     ],
     logo: '/img/plugin-logo.svg',
     editLink: {
-      pattern: 'https://github.com/nystudio107/craft-retour/edit/develop/docs/docs/:path',
+      pattern: 'https://github.com/nystudio107/craft-retour/edit/develop-v4/docs/docs/:path',
       text: 'Edit this page on GitHub'
     },
     algolia: {
@@ -49,7 +49,7 @@ export default defineConfig({
       {
         text: 'v4', items: [
           {text: 'v4', link: '/'},
-          {text: 'v3', link: '/v3/'},
+          {text: 'v3', link: 'https://nystudio107.com/docs/retour/v3/'},
         ],
       },
     ],