+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Retour Changelog
 
+## 4.1.12 - 2023.04.07
+### Changed
+* Enforce full URL matching for short links when used within a multi-site setup and enabling any of the content translation method for the field. ([#277](https://github.com/nystudio107/craft-retour/issues/277))
+
+### Fixed
+* Fixed a bug where short links would not work correctly within Matrix fields. ([#279](https://github.com/nystudio107/craft-retour/issues/279))
+* Fixed a bug where short links would not be cleaned up appropriately when updating a Short Link value on an element. ([#277](https://github.com/nystudio107/craft-retour/issues/277))
+* Fixed a bug where short links would not work correctly with multi-sites. ([#277](https://github.com/nystudio107/craft-retour/issues/277))
+
 ## 4.1.11 - 2023.03.07
 ### Added
 * Added the ability to disable statistics recording entirely via `enableStatistics` in the `config.php` ([#272](https://github.com/nystudio107/craft-retour/pull/272))